max number of particles can be specified. Particles can be generated using different blending styles.

+++ 3DGameEngine/3DGameEngine/res/basicParticlePhy.vert
@@ -1,11 +1,11 @@
 #version 400                                                                        
                                                                                     
-layout (location = 0) in int Type;                                                
+layout (location = 0) in float Type;                                                
 layout (location = 1) in vec3 Position;                                             
 layout (location = 2) in vec3 Velocity;                                             
 layout (location = 3) in float Age;                                                 
                                                                                     
-out int Type0;                                                                    
+out float Type0;                                                                    
 out vec3 Position0;                                                                 
 out vec3 Velocity0;                                                                 
 out float Age0;
