+++ 3DGameEngine/3DGameEngine/res/particleRenderer.vert
@@ -1,9 +1,9 @@
 #version 400
 
-layout (location = 0) in int Type;                                                
+layout (location = 0) in float Type;                                                
 layout (location = 1) in vec3 Position; 
 
-out int Type0;                                             
+out float Type0;                                             
                                                                         
 void main()                                                                         
 {
